log operation as int for fmtlib 8.1.1

+++ cpp/benchmarks/core/BinaryEW.cpp
@@ -116,7 +116,7 @@ static std::function<Tensor(const Tensor&, const Tensor&)> MakeOperation(
             };
 
         default:
-            utility::LogError("Unknown operation {}", op);
+            utility::LogError("Unknown operation {}", static_cast<int>(op));
     }
 }
